// src/types/CardType.ts
export type CardType = {
  answer: string;
  answerImg: string;
  question: string;
  questionImg: string;
  cardsPack_id: string;
  comments: string;
  grade: number;
  more_id: string;
  shots: number;
  user_id: string;
  created: string;
  updated: string;
  _id: string;
  rating: number;
  type: string;
};